Re #13738 Sample temperature, mag field could be a string range
Plan to get sumWorkspaces to put a range in these logs e.g. "1000.0 -
1020.0". In this case printRunInfo needs to deal with this when writing
the values out.

As a side effect this means the values will be written as e.g. "290" or
"290.5", not "290.00000000000000" which is an improvement.

(This is a temporary summed workspace for Muon Analysis only - the string
log value doesn't need to be dealt with elsewhere).